refactor : 기존 score 관련 패키지 이동 model -> score
