+++ src/test/java/bowling/model/score/StateGeneratorTest.java
@@ -1,10 +1,13 @@
-package bowling.model;
+package bowling.model.score;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import bowling.model.Pin;
+import bowling.model.score.ScoreGenerator;
 
 class StateGeneratorTest {
